+++ packages/shared/src/index.ts
@@ -6,5 +6,5 @@ export * from './hooks/index.js'
 export * from './locales/i18n_generated.js'
 export * from './locales/languages.js'
 export * from './types.js'
-export * from './wallet/index.js'
+export * from './UI/wallet/index.js'
 export * from './utils/index.js'
